requirements.txt
certifi==2019.11.28
chardet==3.0.4
docutils==0.16
idna==2.8
#Kivy==1.11.1
#kivy-deps.glew==0.1.12
kivy-deps.sdl2
Kivy-Garden==0.1.4
kivymd==0.103.0
peewee==3.13.1
Pillow==7.0.0
Pygments==2.5.2
#pypiwin32==223
#pywin32==227
requests==2.22.0
urllib3==1.25.8
